Double to int for LEVEL#_PLATFORM_REAR_LIFT_LEVEL

+++ Deep Space 2019/robot/RobotMap.java
@@ -105,8 +105,8 @@ public class RobotMap {
   public static final double BOOMERANG_RETRACTED_POSITION = 1.;
 
   // Rear Lift Level Positions
-  public static final double LEVEL2_PLATFORM_REAR_LIFT_LEVEL = 1.;
-  public static final double LEVEL3_PLATFORM_REAR_LIFT_LEVEL = 1.;
+  public static final int LEVEL2_PLATFORM_REAR_LIFT_LEVEL = 1;
+  public static final int LEVEL3_PLATFORM_REAR_LIFT_LEVEL = 1;
 
   // Collision Detection for NavX
   public static final double COLLISION_THRESHOLD_Y = .5; // in Gs
